Change Timings to a monotonically increasing timer
Timings, as a slice, is fraught with synchronization issues when you
have multiple queries ongoing at the same time. Since it's a slice of
time.Durations, information is lost as to which query each timing is
for. Consequently, the only valuable thing you can do with it is add
them together to come up with a total amount of time for certain actions
(which is done in Reverb). This replaces Timings with an Elapsed value,
which is atomically incremented by the time it takes for queries to
execute. Capturing the Elapsed value before and after taking some action
with Pop will show the amount of time Pop took (think car odometer).