Synchronize access to tableMap
TableMap, as previously used, suffered from a data race, where two
goroutines could try to read and write the same value for a particular
map key. Logically, this is fine, since the racing goroutines would both
ultimately resolve to the same value. Unfortunately, the internals of
maps do not play well with racing goroutines in this way, and it can
cause a runtime panic. This installs a mutex to control access to the
map.