Remove "Connections" singleton
With the singleton you can't use the package without already doing
stuff, e.g. loading and parsing the "database.yml". And if there is
something in "database.yml" POP doesnt like it will die on you with a
panic.

Now the file will be read and parsed when the user tries to make
a connection using "Connect".