Database creation
Hello Mark,

I encountered an issue when i ran `buffalo db create -a` . It displayed the normal message for successful creation of database, but running the application could not create transaction. After checking, I found out non of the database was not created.

Using the create database command with option (-c ~command~)  before the postgres URL solved the issue.

I can't confirm if it's just a windows command prompt problem, but this problem occurred on windows.