Update database template YAML to conform to spec
This commit updates the four database template files to begin using
three hyphens, which per the YAML specification is used to signal the
start of a document.

While not invalid to not include this, as most of the tools don't
require this to properly function, by making this change ensures that
directives are properly parsed in all parsers.

```
3.2.3.4. Directives

Each document may be associated with a set of directives. A directive has a
name and an optional sequence of parameters. Directives are instructions to the
YAML processor, and like all other presentation details are not reflected in
the YAML serialization tree or representation graph. This version of YAML defines
a two directives, “YAML” and “TAG”. All other directives are reserved for future
versions of YAML.
```