Cockroach URL scheme is now postgres
Before this patch, when using a raw URL connection string such as
`crdb://foo/baz`, the cockroach dialect would not replace it with
`postgres://`. This causes an issue where the `lib/pq` driver
completely ignores the URL and simply falls back to the defaults
(localhost, 5432).

This patch changes that, by replacing `cockroach` (and aliases)
with `postgres` when returning the URL.

Signed-off-by: aeneasr <aeneas@ory.sh>