fix: remove ineffective and bug-ridden model `tableMap` cache
This patch removes the Model's `tableMap`. The cache seems to have served a purpose before `TableName()` was introduced. Now however, it does not significantly reduce computation time and instead introduces a myriad of really difficult to debug problems. This usually happens due to cacheKey collisions. The amount of time the TCP connection will use to query the database is significantly more than calling `reflect.Elem()`, further delegitimizing the complexity of having a cache here.

This patch completely removes the cache and adds failing test cases to ensure that future regressions do not resurface.

Signed-off-by: aeneasr <3372410+aeneasr@users.noreply.github.com>