Avoid dumping sqlite system tables into schema
sqlite includes some of its system tables in the .schema output by
default, which means it cannot be imported back directly into sqlite (as
LoadSchema attempts to do), as you're not allowed to recreate these
tables.

The -nosys option can be used to prevent this behaviour.

Particular issues have been seen with the sqlite_sequence table, which
appears if you use an AUTOINCREMENT column as an ID for example.