slices: update Int.Scan() to parse empty slices
Previously, Int.Scan() returned the slice '[0]' when reading an empty
slice of ints from the database. This is a bug, as the output should
also be empty. This patch fixes the bug and adds regression tests.
Additionally, the code is updated to percolate up the errors from
integer to string conversion, which was previously swallowed.